refactor: settings and handlers for form interactions
