+++ .gitignore
@@ -6,3 +6,4 @@ build
 ignore
 .idea
 .DS_Store
+__pycache__
