+++ worksafe_nop/data.json
@@ -3,6 +3,7 @@
   "LAST_NAME": "Doe",
   "PHONE_NUMBER": "1234567890",
   "EMAIL": "john.doe@example.com",
+
   "PROJECT_TYPE": "Asbestos",
   "ADDRESS": "123 Main Street",
   "ADDRESS_DETAILS": "Building entrance on north side",
@@ -12,14 +13,18 @@
   "START_TIME": "08:00",
   "END_TIME": "16:00",
   "WORKER_COUNT": "3",
+
   "WORKSAFE_BC_ACCOUNT_NUMBER": "123456",
   "EMPLOYER_NAME": "ABC Construction",
   "EMPLOYER_ADDRESS": "456 1st Avenue",
+
   "ASBESTOS_CHECK": true,
   "LEAD_CHECK": false,
   "OTHER_CHECK": false,
+  "ASBESTOS_LICENSE_NUMBER": "12345678",
   "IN_CHARGE": "SubmitterinCharge",
   "JOB_TITLE": "Project Manager",
+  "ASBESTOS_CERTIFICATION_NUMBER": "12345678",
   "MULTIPLE_EMPLOYERS": "no",
   "OTHER_CONSULTING_FIRMS": "XYZ Engineering (Safety Consultant)",
